requirements.txt
# Agent orchestration framework
crewai==0.41.1
crewai-tools==0.4.26

# LLM integrations (required by crewai)
langchain==0.2.14
langchain-community==0.2.12
langchain-google-genai==1.0.8

# FastAPI for API service
fastapi==0.111.0
uvicorn==0.30.0
pydantic>=2.6.1,<3.0.0
python-multipart==0.0.9

# Background tasks and queuing
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.31
alembic>=1.13.1

# Document processing
python-docx==1.1.2
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.17.0
Pillow>=10.0.0

# Google AI
google-generativeai==0.7.2

# HTTP client
httpx==0.27.0
aiohttp==3.9.5

# Data handling
pandas==2.1.3
numpy==1.25.2
pydantic-settings==2.1.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest>=8.0.0
pytest-asyncio==0.21.1

# Environment and config
python-dotenv==1.0.0

# JSON Schema validation and extraction
jsonschema>=4.22.0
openai>=1.40.0
anthropic>=0.31.0
